+++ _example/simple-echo/main.go
@@ -4,16 +4,21 @@ import (
 	"fmt"
 
 	prompt "github.com/elk-language/go-prompt"
+	pstrings "github.com/elk-language/go-prompt/strings"
 )
 
-func completer(in prompt.Document) []prompt.Suggest {
+func completer(in prompt.Document) ([]prompt.Suggest, pstrings.RuneNumber, pstrings.RuneNumber) {
 	s := []prompt.Suggest{
 		{Text: "users", Description: "Store the username and age"},
 		{Text: "articles", Description: "Store the article text posted by user"},
 		{Text: "comments", Description: "Store the text commented to articles"},
 		{Text: "groups", Description: "Combine users with specific rules"},
 	}
-	return prompt.FilterHasPrefix(s, in.GetWordBeforeCursor(), true)
+	endIndex := in.CurrentRuneIndex()
+	w := in.GetWordBeforeCursor()
+	startIndex := endIndex - pstrings.RuneCountInString(w)
+
+	return prompt.FilterHasPrefix(s, w, true), startIndex, endIndex
 }
 
 func main() {
@@ -22,7 +27,6 @@ func main() {
 		prompt.WithTitle("sql-prompt"),
 		prompt.WithHistory([]string{"SELECT * FROM users;"}),
 		prompt.WithPrefixTextColor(prompt.Yellow),
-		prompt.WithPreviewSuggestionTextColor(prompt.Blue),
 		prompt.WithSelectedSuggestionBGColor(prompt.LightGray),
 		prompt.WithSuggestionBGColor(prompt.DarkGray),
 		prompt.WithCompleter(completer),